patron state a base datos
