+++ src/main/java/app/persistence/conection/BaseDatosSql.java
@@ -1,4 +1,4 @@
-package app.persistencia.Conection;
+package app.persistence.conection;
 
 import javax.persistence.EntityManagerFactory;
