+++ src/main/java/app/persistence/conection/MySqlJorge.java
@@ -1,4 +1,4 @@
-package app.persistencia.Conection;
+package app.persistence.conection;
 
 import javax.persistence.EntityManagerFactory;
 import javax.persistence.Persistence;
@@ -8,24 +8,22 @@ import java.util.Map;
 /**
  * Created by gabriellorenzatti on 10/6/17.
  */
-public class MySqlGabi extends ConeccionMySql {
+public class MySqlJorge extends ConeccionMySql {
 
 
-    private String conexionGabi = "jdbc:mysql://localhost:3306/pupi";
-    private String userGabi = "root";
-    private String passwordGabi = "6732";
+    private String conexionJorge = "jdbc:mysql://localhost:3306/pupi";
+    private String userJorge = "root";
+    private String passwordJorge = "mypassword";
     @Override
     public EntityManagerFactory executeAction(EntityManagerFactory emf) {
-        System.out.println("algo");
-        System.out.println("============================= CONFIGURO local MYSQL gabi");
+        System.out.println("============================= CONFIGURO local MYSQL jorge");
         Map<String, String> persistenceMap = new HashMap<>();
-        persistenceMap.put("javax.persistence.jdbc.url", conexionGabi);
-        persistenceMap.put("javax.persistence.jdbc.user", userGabi);
-        persistenceMap.put("javax.persistence.jdbc.password", passwordGabi);
+        persistenceMap.put("javax.persistence.jdbc.url", conexionJorge);
+        persistenceMap.put("javax.persistence.jdbc.user", userJorge);
+        persistenceMap.put("javax.persistence.jdbc.password", passwordJorge);
         persistenceMap.put("javax.persistence.jdbc.driver", "com.mysql.jdbc.Driver");
         persistenceMap.put("javax.persistence.schema-generation.database.action", "create-or-extend-tables");
         emf = Persistence.createEntityManagerFactory("PersistenceUnit", persistenceMap);
         return emf;
     }
-
 }
