+++ src/main/java/app/persistence/conection/MySqlJose.java
@@ -1,4 +1,4 @@
-package app.persistencia.Conection;
+package app.persistence.conection;
 
 import javax.persistence.EntityManagerFactory;
 import javax.persistence.Persistence;
@@ -8,25 +8,25 @@ import java.util.Map;
 /**
  * Created by gabriellorenzatti on 10/6/17.
  */
-public class MySqlFede extends ConeccionMySql {
+public class MySqlJose extends ConeccionMySql {
 
 
-    private String conexionFede = "jdbc:mysql://localhost:3306/pupi";
-    private String userFede = "root";
-    private String passwordFede = "mypassword";
-
+    private String conexionJose = "jdbc:mysql://localhost:6603/pupi";
+    private String userJose = "root";
+    private String passwordJose = "mypassword";
     @Override
     public EntityManagerFactory executeAction(EntityManagerFactory emf) {
-        System.out.println("============================= CONFIGURO local MYSQL fede");
+        System.out.println("algo");
+        System.out.println("============================= CONFIGURO local MYSQL jose");
         Map<String, String> persistenceMap = new HashMap<>();
-        persistenceMap.put("javax.persistence.jdbc.url", conexionFede);
-        persistenceMap.put("javax.persistence.jdbc.user", userFede);
-        persistenceMap.put("javax.persistence.jdbc.password", passwordFede);
-
+        persistenceMap.put("javax.persistence.jdbc.url", conexionJose);
+        persistenceMap.put("javax.persistence.jdbc.user", userJose);
+        persistenceMap.put("javax.persistence.jdbc.password", passwordJose);
         persistenceMap.put("javax.persistence.jdbc.driver", "com.mysql.jdbc.Driver");
         persistenceMap.put("javax.persistence.schema-generation.database.action", "create-or-extend-tables");
         emf = Persistence.createEntityManagerFactory("PersistenceUnit", persistenceMap);
         return emf;
     }
+
 
 }
