+++ src/main/java/app/persistence/conection/MySqlPaolo.java
@@ -1,4 +1,4 @@
-package app.persistencia.Conection;
+package app.persistence.conection;
 
 import javax.persistence.EntityManagerFactory;
 import javax.persistence.Persistence;
@@ -8,25 +8,22 @@ import java.util.Map;
 /**
  * Created by gabriellorenzatti on 10/6/17.
  */
-public class MySqlJose extends ConeccionMySql {
+public class MySqlPaolo extends ConeccionMySql {
 
-
-    private String conexionJose = "jdbc:mysql://localhost:6603/pupi";
-    private String userJose = "root";
-    private String passwordJose = "mypassword";
+    private String conexionPaolo = "jdbc:mysql://localhost:3306/pupi";
+    private String userPaolo = "root";
+    private String passwordPaolo = "MilikiJimenezCrack77";
     @Override
     public EntityManagerFactory executeAction(EntityManagerFactory emf) {
-        System.out.println("algo");
         System.out.println("============================= CONFIGURO local MYSQL jose");
         Map<String, String> persistenceMap = new HashMap<>();
-        persistenceMap.put("javax.persistence.jdbc.url", conexionJose);
-        persistenceMap.put("javax.persistence.jdbc.user", userJose);
-        persistenceMap.put("javax.persistence.jdbc.password", passwordJose);
+        persistenceMap.put("javax.persistence.jdbc.url", conexionPaolo);
+        persistenceMap.put("javax.persistence.jdbc.user", userPaolo);
+        persistenceMap.put("javax.persistence.jdbc.password", passwordPaolo);
         persistenceMap.put("javax.persistence.jdbc.driver", "com.mysql.jdbc.Driver");
         persistenceMap.put("javax.persistence.schema-generation.database.action", "create-or-extend-tables");
         emf = Persistence.createEntityManagerFactory("PersistenceUnit", persistenceMap);
         return emf;
     }
-
 
 }
